--- src/test/routes/track.me.constants.json
@@ -1,11 +1,0 @@
-{
-	"jwtTestUser": {
-		"id": 24
-	},
-	"initialTrack": {
-		"name": "Despacito",
-		"albumId": 240,
-		"artists": [2, 1]
-	},
-	"validTrackId": 1
-}
